split file by build platform
